rehauled app to make sense
