+++ src/components/Display.js
@@ -1,91 +1,15 @@
 import React from "react";
-import Dashboard from "./Dashboard";
 
-class Display extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-      balls: 0,
-      strikes: 0,
-      fouls: 0,
-      hits: 0
-    };
-  }
-
-  incrementBalls = () => {
-    if (this.state.balls < 3) {
-      this.setState({ ...this.state, balls: this.state.balls + 1 });
-    } else {
-      this.setState({
-        balls: 0,
-        strikes: 0,
-        fouls: 0
-      });
-    }
-  };
-
-  incrementStrikes = () => {
-    if (this.state.strikes < 2) {
-      this.setState({
-        ...this.state,
-        strikes: this.state.strikes + 1
-      });
-    } else {
-      this.setState({
-        balls: 0,
-        strikes: 0,
-        fouls: 0
-      });
-    }
-  };
-
-  incrementHits = () => {
-    this.setState({
-      balls: 0,
-      strikes: 0,
-      fouls: 0,
-      hits: this.state.hits + 1
-    });
-  };
-
-  incrementFouls = () => {
-    if (this.state.strikes < 2) {
-      this.setState({
-        // ...this.state,
-        strikes: this.state.strikes + 1
-      });
-    } else {
-      this.setState({
-        // ...this.state
-      });
-    }
-    this.setState({ fouls: this.state.fouls + 1 });
-  };
-
-  render() {
-    return (
-      <>
-        <Dashboard
-          balls={this.state.balls}
-          strikes={this.state.strikes}
-          fouls={this.state.fouls}
-          hits={this.state.hits}
-        />
-        <button title="ballButton" onClick={this.incrementBalls}>
-          Ball
-        </button>
-        <button title="strikeButton" onClick={this.incrementStrikes}>
-          Strike
-        </button>
-        <button title="foulButton" onClick={this.incrementFouls}>
-          Foul
-        </button>
-        <button title="hitButton" onClick={this.incrementHits}>
-          Hit
-        </button>
-      </>
-    );
-  }
+function Display(props) {
+  console.log(props.fouls);
+  return (
+    <>
+      <h2 title="ballCount">Balls:{props.balls}</h2>
+      <h2 title="strikeCount">Strikes:{props.strikes}</h2>
+      <h2 title="foulCount">Fouls:{props.fouls}</h2>
+      <h2 title="hitCount">Hits:{props.hits}</h2>
+    </>
+  );
 }
 
 export default Display;
